refact: setting.tsx 구조 개선
