+++ src/features/settings/index.ts
@@ -1,0 +1,12 @@
+/**
+ * 설정 기능 내보내기
+ */
+
+// 컴포넌트
+export { default as Settings } from "./components/Settings";
+
+// 훅
+export { useSettings } from "./hooks/useSettings";
+
+// 타입
+export * from "./types";
